refactor: filters change the url
