+++ app/views/partials/_sort-drop-down.html.erb
@@ -1,3 +1,3 @@
-<select name="sort_by" id="sort_by" class="bg-white form-select h-full py-0 pl-3 pr-8 shadow-sm bg-transparent text-gray-500 sm:text-sm sm:leading-5" @input="window.location.href = event.target.value">
+<select name="sort_by" id="sort_by" class="w-full sm:w-auto text-sm text-gray-700 bg-white form-select shadow-sm text-gray-500 sm:text-sm sm:leading-5" @input="window.location.href = event.target.value">
   <%= yield %>
 </select>
