+++ app/views/projects/_categories-nav.html.erb
@@ -1,4 +1,4 @@
-<div class="-mt-6 pb-4 w-full flex justify-center border-b border-gray-200">
+<div class="-mt-6 pb-4 w-full flex flex-col sm:flex-row justify-center border-b border-gray-200">
   <%
     classes = 'text-sm mx-4'
     active_classes = 'font-bold border-b-2 border-indigo-600'
