+++ app/views/projects/_category-info.html.erb
@@ -1,5 +1,27 @@
-<div class="mx-auto w-1/4 text-center mt-8">
-  <div class="inline-block rounded-full bg-white fill-current w-24 h-24 flex justify-center items-center mx-auto my-4 bg-indigo-600 fill-current text-white"><%= inline_svg_pack_tag "media/svgs/category-#{@project_category[:name].downcase}.svg", class: 'fill-current text-white' %></div>
-  <div class="font-bold text-lg mb-4"><%= @project_category[:name] %></div>
-  <div class="mb-4 leading-tight"><%= @project_category[:body] %></div>
+<%
+  if @project_category.blank?
+    @project_category = {
+      name: 'All Categories',
+      color: 'indigo',
+    }
+  end
+
+  # We need to explicitly write the full color instead of interpolate it so we can later safely purge the unsued css
+  case @project_category[:color]
+  when 'indigo'
+    bg_color = 'bg-indigo-600'
+  when 'orange'
+    bg_color = 'bg-orange-600'
+  when 'blue'
+    bg_color = 'bg-blue-600'
+  when 'purple'
+    bg_color = 'bg-purple-600'
+  else
+    bg_color = 'bg-indigo-600'
+  end
+%>
+<div class="mx-auto w-1/3 text-center mt-8">
+  <div class="inline-block rounded-full bg-white fill-current w-24 h-24 flex justify-center items-center mx-auto my-4 fill-current text-white <%= bg_color %>"><%= inline_svg_pack_tag "media/svgs/category-#{@project_category[:name].parameterize.downcase}.svg", class: 'fill-current text-white' %></div>
+  <div class="font-bold text-2xl mb-4 text-gray-700"><%= @project_category[:name] %></div>
+  <div class="mb-4 leading-tight text-gray-700"><%= @project_category[:body] %></div>
 </div>
