+++ app/views/projects/_filter-bar-filter.html.erb
@@ -1,7 +1,6 @@
-<div class="relative bg-white py-2 px-4 mx-2" x-data="dropdown<%= filter.camelize %>()" @click.away="open = false">
-  <div class="flex text-sm items-center select-none" @click="open = !open">
-    <div class="cursor-pointer"><span x-text="selectedCount()"></span> <%= label %></div>
-    <%= inline_svg_pack_tag('media/svgs/chevron-right.svg', class: 'h-4 fill-current text-gray-600 transform rotate-90') %>
+<div class="relative" x-data="dropdown<%= filter.camelize %>()" @click.away="open = false" data-turbolinks-permanent>
+  <div class="flex text-sm text-gray-700 items-center justify-between select-none form-select pr-8 bg-white border-gray-300 rounded shadow-sm" @click="open = !open">
+    <div class="cursor-pointer flex-shrink-0"><span x-text="selectedCount()"></span> <span x-text="Covid.pluralize('<%= label %>', selected.length)"></span></div>
   </div>
   <div class="absolute left-0 mt-4 p-4 bg-white z-30 shadow-lg w-48 flex flex-col" x-show="open">
     <div class="overflow-auto max-h-400px">
@@ -25,9 +24,7 @@
       open: false,
       options: <%= options.to_json.html_safe %>,
       applyFilters() {
-        const url = `/projects?<%= filter %>=${this.selected.map(param => encodeURIComponent(param)).join(',')}`
-        console.log('applyFilters', (url))
-        window.location.href = url
+        Covid.applyFiltersAndGo('<%= filter %>', this.selected)
       },
       resetFilters() {
         this.selected = []
