+++ package.json
@@ -9,9 +9,11 @@
     "@tailwindcss/ui": "^0.1.3",
     "jquery": "^3.4.1",
     "js-cookie": "^2.2.1",
+    "pluralize": "^8.0.0",
     "select2": "^4.0.13",
     "tailwindcss": "^1.2.0",
     "turbolinks": "^5.2.0",
+    "urijs": "^1.19.2",
     "yarn": "^1.22.4"
   },
   "version": "0.1.0",
